debug metadata validation script and add debug statements to main script, revert to original script names without _new
